backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
alembic
psycopg2-binary
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart